ceph.api/requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
rgwadmin==2.3.1
flask-cors
